// source/key.enum.ts
export const enum Key {
  ALT = 18,
  BACKSPACE = 8,
  CONTROL = 17,
  DOWN_ARROW = 40,
  END = 35,
  ENTER = 13,
  ESCAPE = 27,
  HOME = 36,
  LEFT_ARROW = 37,
  NUMBER_0 = 48,
  NUMBER_1 = 49,
  NUMBER_2 = 50,
  NUMBER_3 = 51,
  NUMBER_4 = 52,
  NUMBER_5 = 53,
  NUMBER_6 = 54,
  NUMBER_7 = 55,
  NUMBER_8 = 56,
  NUMBER_9 = 57,
  PAGE_DOWN = 34,
  PAGE_UP = 33,
  RIGHT_ARROW = 39,
  SHIFT = 16,
  SPACEBAR = 32,
  TAB = 9,
  UP_ARROW = 38,
}